Split Audio/VideoOutputStream source
Summary: Refactor Audio/VideoOutputStream.

Reviewed By: nateanl

Differential Revision: D43613008

fbshipit-source-id: 6956e1501b6efa2160ad35e5a743d04daf786e05
