+++ torchaudio/csrc/ffmpeg/CMakeLists.txt
@@ -18,6 +18,8 @@ set(
   stream_reader/stream_reader.cpp
   stream_writer/encoder.cpp
   stream_writer/output_stream.cpp
+  stream_writer/audio_output_stream.cpp
+  stream_writer/video_output_stream.cpp
   stream_writer/stream_writer.cpp
   compat.cpp
   utils.cpp
